[CHK-1921]: refactor(v2): handle different event versions (#147)
* fix(event parsing): invalid event written to dead-letter queue

* fix: spotless

* feat(v2): move event handlers to separate v1 package

* refactor: add common queue consumer with version switching

* Chk 1921 v2 event deserialization (#151)

* fix(queue-event): fix event deserialization on different versions by discriminator

* fix(queue-event): parsing old events

* fix(lint): spotless

* merge

* fix: update commons version

---------
