+++ src/main/kotlin/it/pagopa/ecommerce/eventdispatcher/exceptions/InvalidEventException.kt
@@ -5,5 +5,5 @@ import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
 
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-class InvalidEventException(eventPayload: ByteArray) :
-  RuntimeException("Invalid input event: ${eventPayload.toString(StandardCharsets.UTF_8)}")
+class InvalidEventException(eventPayload: ByteArray, cause: Throwable?) :
+  RuntimeException("Invalid input event: ${eventPayload.toString(StandardCharsets.UTF_8)}", cause)
