+++ src/main/kotlin/it/pagopa/ecommerce/eventdispatcher/exceptions/TransactionEventsPreconditionsNotMatchedException.kt
@@ -1,7 +1,7 @@
 package it.pagopa.ecommerce.eventdispatcher.exceptions
 
+import it.pagopa.ecommerce.commons.domain.TransactionId
 import it.pagopa.ecommerce.commons.domain.v1.TransactionEventCode
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId
 
 class TransactionEventsPreconditionsNotMatchedException(
   transactionId: TransactionId,
