+++ src/main/kotlin/it/pagopa/ecommerce/eventdispatcher/repositories/TransactionsEventStoreRepository.kt
@@ -1,10 +1,12 @@
 package it.pagopa.ecommerce.eventdispatcher.repositories
 
-import it.pagopa.ecommerce.commons.documents.v1.TransactionEvent
+import it.pagopa.ecommerce.commons.documents.BaseTransactionEvent
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Flux
 
 interface TransactionsEventStoreRepository<T> :
-  ReactiveCrudRepository<TransactionEvent<T>, String> {
-  fun findByTransactionIdOrderByCreationDateAsc(idTransaction: String): Flux<TransactionEvent<T>>
+  ReactiveCrudRepository<BaseTransactionEvent<T>, String> {
+  fun findByTransactionIdOrderByCreationDateAsc(
+    idTransaction: String
+  ): Flux<BaseTransactionEvent<T>>
 }
