--- src/main/resources/logback-spring.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <springProperty name="ECS_SERVICE_VERSION" source="build.version"/>
-    <appender name="ECS_JSON_CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder class="co.elastic.logging.logback.EcsEncoder">
-            <serviceName>${ECS_SERVICE_NAME:-undefined}</serviceName>
-            <serviceVersion>${ECS_SERVICE_VERSION}</serviceVersion>
-            <serviceEnvironment>${ECS_SERVICE_ENVIRONMENT:-undefined}</serviceEnvironment>
-        </encoder>
-    </appender>
-    <root level="INFO">
-        <appender-ref ref="ECS_JSON_CONSOLE"/>
-    </root>
-</configuration>
