+++ src/test/kotlin/it/pagopa/ecommerce/eventdispatcher/client/NodeClientTest.kt
@@ -1,6 +1,6 @@
 package it.pagopa.ecommerce.eventdispatcher.client
 
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId
+import it.pagopa.ecommerce.commons.domain.TransactionId
 import it.pagopa.ecommerce.commons.v1.TransactionTestUtils
 import it.pagopa.ecommerce.eventdispatcher.exceptions.BadClosePaymentRequest
 import it.pagopa.ecommerce.eventdispatcher.exceptions.BadGatewayException
