+++ src/test/kotlin/it/pagopa/ecommerce/eventdispatcher/queues/v1/TransactionNotificationsQueueConsumerTest.kt
@@ -1,12 +1,12 @@
-package it.pagopa.ecommerce.eventdispatcher.queues
+package it.pagopa.ecommerce.eventdispatcher.queues.v1
 
 import com.azure.core.util.BinaryData
 import com.azure.core.util.serializer.TypeReference
 import com.azure.spring.messaging.checkpoint.Checkpointer
 import com.azure.storage.queue.QueueAsyncClient
 import it.pagopa.ecommerce.commons.documents.v1.*
+import it.pagopa.ecommerce.commons.domain.Email
 import it.pagopa.ecommerce.commons.domain.v1.TransactionEventCode
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId
 import it.pagopa.ecommerce.commons.domain.v1.pojos.BaseTransactionWithRequestedUserReceipt
 import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto
 import it.pagopa.ecommerce.commons.queues.QueueEvent
@@ -15,15 +15,13 @@ import it.pagopa.ecommerce.commons.queues.TracingUtilsTests
 import it.pagopa.ecommerce.commons.v1.TransactionTestUtils.*
 import it.pagopa.ecommerce.eventdispatcher.client.NotificationsServiceClient
 import it.pagopa.ecommerce.eventdispatcher.client.PaymentGatewayClient
-import it.pagopa.ecommerce.eventdispatcher.exceptions.NoRetryAttemptsLeftException
 import it.pagopa.ecommerce.eventdispatcher.repositories.TransactionsEventStoreRepository
 import it.pagopa.ecommerce.eventdispatcher.repositories.TransactionsViewRepository
-import it.pagopa.ecommerce.eventdispatcher.services.eventretry.NotificationRetryService
-import it.pagopa.ecommerce.eventdispatcher.services.eventretry.RefundRetryService
-import it.pagopa.ecommerce.eventdispatcher.utils.DEAD_LETTER_QUEUE_TTL_SECONDS
-import it.pagopa.ecommerce.eventdispatcher.utils.UserReceiptMailBuilder
-import it.pagopa.ecommerce.eventdispatcher.utils.queueSuccessfulResponse
+import it.pagopa.ecommerce.eventdispatcher.services.eventretry.v1.NotificationRetryService
+import it.pagopa.ecommerce.eventdispatcher.services.eventretry.v1.RefundRetryService
+import it.pagopa.ecommerce.eventdispatcher.utils.*
 import it.pagopa.generated.ecommerce.gateway.v1.dto.VposDeleteResponseDto
+import it.pagopa.generated.notifications.templates.success.SuccessTemplate
 import it.pagopa.generated.notifications.v1.dto.NotificationEmailRequestDto
 import it.pagopa.generated.notifications.v1.dto.NotificationEmailResponseDto
 import java.nio.charset.StandardCharsets
@@ -47,17 +45,12 @@ import reactor.test.StepVerifier
 
 @OptIn(ExperimentalCoroutinesApi::class)
 @ExtendWith(MockitoExtension::class)
-class TransactionNotificationsRetryQueueConsumerTest {
-  private val paymentGatewayClient: PaymentGatewayClient = mock()
+class TransactionNotificationsQueueConsumerTest {
 
   private val transactionsEventStoreRepository: TransactionsEventStoreRepository<Any> = mock()
 
   private val transactionUserReceiptRepository:
     TransactionsEventStoreRepository<TransactionUserReceiptData> =
-    mock()
-
-  private val transactionRefundRepository:
-    TransactionsEventStoreRepository<TransactionRefundedData> =
     mock()
 
   private val transactionsViewRepository: TransactionsViewRepository = mock()
@@ -66,19 +59,18 @@ class TransactionNotificationsRetryQueueConsumerTest {
 
   private val notificationRetryService: NotificationRetryService = mock()
 
-  private val refundRetryService: RefundRetryService = mock()
-
   private val notificationsServiceClient: NotificationsServiceClient = mock()
 
   private val userReceiptMailBuilder: UserReceiptMailBuilder = mock()
+  private val transactionRefundRepository:
+    TransactionsEventStoreRepository<TransactionRefundedData> =
+    mock()
+  private val paymentGatewayClient: PaymentGatewayClient = mock()
+  private val refundRetryService: RefundRetryService = mock()
 
   private val tracingUtils = TracingUtilsTests.getMock()
 
   @Captor private lateinit var transactionViewRepositoryCaptor: ArgumentCaptor<Transaction>
-
-  @Captor
-  private lateinit var transactionRefundEventStoreCaptor:
-    ArgumentCaptor<TransactionEvent<TransactionRefundedData>>
 
   @Captor
   private lateinit var transactionUserReceiptCaptor:
@@ -86,41 +78,45 @@ class TransactionNotificationsRetryQueueConsumerTest {
 
   @Captor private lateinit var retryCountCaptor: ArgumentCaptor<Int>
 
+  @Captor
+  private lateinit var transactionRefundEventStoreCaptor:
+    ArgumentCaptor<TransactionEvent<TransactionRefundedData>>
+
   private val deadLetterQueueAsyncClient: QueueAsyncClient = mock()
 
   private val transactionNotificationsRetryQueueConsumer =
-    TransactionNotificationsRetryQueueConsumer(
+    TransactionNotificationsQueueConsumer(
       transactionsEventStoreRepository = transactionsEventStoreRepository,
       transactionUserReceiptRepository = transactionUserReceiptRepository,
       transactionsViewRepository = transactionsViewRepository,
       notificationRetryService = notificationRetryService,
+      userReceiptMailBuilder = userReceiptMailBuilder,
+      notificationsServiceClient = notificationsServiceClient,
       transactionsRefundedEventStoreRepository = transactionRefundRepository,
       paymentGatewayClient = paymentGatewayClient,
       refundRetryService = refundRetryService,
-      userReceiptMailBuilder = userReceiptMailBuilder,
-      notificationsServiceClient = notificationsServiceClient,
       deadLetterQueueAsyncClient = deadLetterQueueAsyncClient,
       deadLetterTTLSeconds = DEAD_LETTER_QUEUE_TTL_SECONDS,
       tracingUtils = tracingUtils)
 
   @Test
-  fun `Should successfully retry send user email for send payment result outcome OK`() = runTest {
+  fun `Should successfully send user email for send payment result outcome OK`() = runTest {
     val transactionUserReceiptData =
       transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-    val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+    val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
     val events =
       listOf(
         transactionActivateEvent(),
         transactionAuthorizationRequestedEvent(),
         transactionAuthorizationCompletedEvent(),
         transactionClosedEvent(TransactionClosureData.Outcome.OK),
-        transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-        notificationErrorEvent)
+        notificationRequested)
         as List<TransactionEvent<Any>>
     val baseTransaction =
       reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
     val transactionId = TRANSACTION_ID
-    val document = transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
+    val document =
+      transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
     Hooks.onOperatorDebug()
     given(checkpointer.success()).willReturn(Mono.empty())
     given(transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
@@ -140,195 +136,48 @@ class TransactionNotificationsRetryQueueConsumerTest {
 
     StepVerifier.create(
         transactionNotificationsRetryQueueConsumer.messageReceiver(
-          BinaryData.fromObject(QueueEvent(notificationErrorEvent, MOCK_TRACING_INFO)).toBytes(),
+          BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
           checkpointer))
       .expectNext()
       .verifyComplete()
     verify(checkpointer, times(1)).success()
     verify(transactionsEventStoreRepository, times(1))
       .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
-    verify(transactionUserReceiptRepository, times(1)).save(any())
     verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
     verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-    verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
     verify(transactionsViewRepository, times(1)).save(any())
+    verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
     verify(transactionRefundRepository, times(0)).save(any())
     verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
+    verify(transactionUserReceiptRepository, times(1)).save(any())
     verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
     assertEquals(TransactionStatusDto.NOTIFIED_OK, transactionViewRepositoryCaptor.value.status)
     val savedEvent = transactionUserReceiptCaptor.value
-    assertEquals(TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT, savedEvent.eventCode)
+    assertEquals(
+      TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT,
+      TransactionEventCode.valueOf(savedEvent.eventCode))
     assertEquals(transactionUserReceiptData, savedEvent.data)
   }
 
   @Test
-  fun `Should successfully retry send user email for send payment result outcome KO`() = runTest {
-    val transactionUserReceiptData =
-      transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-    val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-    val events =
-      listOf(
-        transactionActivateEvent(),
-        transactionAuthorizationRequestedEvent(),
-        transactionAuthorizationCompletedEvent(),
-        transactionClosedEvent(TransactionClosureData.Outcome.OK),
-        transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-        notificationErrorEvent)
-        as List<TransactionEvent<Any>>
-    val baseTransaction =
-      reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-    val transactionId = TRANSACTION_ID
-    val document = transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-    Hooks.onOperatorDebug()
-    given(checkpointer.success()).willReturn(Mono.empty())
-    given(transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
-      .willReturn(Flux.fromIterable(events))
-    given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-      .willReturn(NotificationEmailRequestDto())
-    given(notificationsServiceClient.sendNotificationEmail(any()))
-      .willReturn(Mono.just(NotificationEmailResponseDto().apply { outcome = "OK" }))
-    given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-      .willReturn(Mono.just(document))
-    given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-      Mono.just(it.arguments[0])
-    }
-    given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor))).willAnswer {
-      Mono.just(it.arguments[0])
-    }
-    given(transactionRefundRepository.save(capture(transactionRefundEventStoreCaptor))).willAnswer {
-      Mono.just(it.arguments[0])
-    }
-    given(paymentGatewayClient.requestVPosRefund(any()))
-      .willReturn(
-        Mono.just(VposDeleteResponseDto().status(VposDeleteResponseDto.StatusEnum.CANCELLED)))
-
-    given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-      .willReturnConsecutively(
-        listOf(
-          Mono.just(
-            transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())),
-          Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
-          Mono.just(
-            transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
-    StepVerifier.create(
-        transactionNotificationsRetryQueueConsumer.messageReceiver(
-          BinaryData.fromObject(QueueEvent(notificationErrorEvent, MOCK_TRACING_INFO)).toBytes(),
-          checkpointer))
-      .expectNext()
-      .verifyComplete()
-    verify(checkpointer, times(1)).success()
-    verify(transactionsEventStoreRepository, times(1))
-      .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
-    verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-    verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-    verify(transactionsViewRepository, times(3)).save(any())
-    verify(transactionRefundRepository, times(2)).save(any())
-    verify(paymentGatewayClient, times(1)).requestVPosRefund(any())
-    verify(transactionUserReceiptRepository, times(1)).save(any())
-    verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-    verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-    val savedEvent = transactionUserReceiptCaptor.value
-    assertEquals(TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT, savedEvent.eventCode)
-    assertEquals(transactionUserReceiptData, savedEvent.data)
-    val expectedStatuses =
-      listOf(
-        TransactionStatusDto.NOTIFIED_KO,
-        TransactionStatusDto.REFUND_REQUESTED,
-        TransactionStatusDto.REFUNDED)
-    val expectedEventCodes =
-      listOf(
-        TransactionEventCode.TRANSACTION_REFUND_REQUESTED_EVENT,
-        TransactionEventCode.TRANSACTION_REFUNDED_EVENT)
-    expectedEventCodes.forEachIndexed { index, eventCode ->
-      assertEquals(eventCode, transactionRefundEventStoreCaptor.allValues[index].eventCode)
-      assertEquals(
-        TransactionStatusDto.NOTIFICATION_ERROR,
-        transactionRefundEventStoreCaptor.allValues[index].data.statusBeforeRefunded)
-    }
-    expectedStatuses.forEachIndexed { index, transactionStatus ->
-      assertEquals(transactionStatus, transactionViewRepositoryCaptor.allValues[index].status)
-    }
-  }
-
-  @Test
-  fun `Should successfully retry send user email for send payment result outcome OK with legacy event`() =
-    runTest {
-      val transactionUserReceiptData =
-        transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-      val events =
-        listOf(
-          transactionActivateEvent(),
-          transactionAuthorizationRequestedEvent(),
-          transactionAuthorizationCompletedEvent(),
-          transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent)
-          as List<TransactionEvent<Any>>
-      val baseTransaction =
-        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-      val transactionId = TRANSACTION_ID
-      val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-      Hooks.onOperatorDebug()
-      given(checkpointer.success()).willReturn(Mono.empty())
-      given(
-          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
-        .willReturn(Flux.fromIterable(events))
-      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-        .willReturn(NotificationEmailRequestDto())
-      given(notificationsServiceClient.sendNotificationEmail(any()))
-        .willReturn(Mono.just(NotificationEmailResponseDto().apply { outcome = "OK" }))
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturn(Mono.just(document))
-      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-        Mono.just(it.arguments[0])
-      }
-      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-
-      StepVerifier.create(
-          transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(notificationErrorEvent).toBytes(), checkpointer))
-        .expectNext()
-        .verifyComplete()
-      verify(checkpointer, times(1)).success()
-      verify(transactionsEventStoreRepository, times(1))
-        .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
-      verify(transactionUserReceiptRepository, times(1)).save(any())
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(1)).save(any())
-      verify(transactionRefundRepository, times(0)).save(any())
-      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      assertEquals(TransactionStatusDto.NOTIFIED_OK, transactionViewRepositoryCaptor.value.status)
-      val savedEvent = transactionUserReceiptCaptor.value
-      assertEquals(TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT, savedEvent.eventCode)
-      assertEquals(transactionUserReceiptData, savedEvent.data)
-    }
-
-  @Test
-  fun `Should successfully retry send user email for send payment result outcome KO with legacy event`() =
+  fun `Should successfully send user email for send payment result outcome KO performing refund for transaction`() =
     runTest {
       val transactionUserReceiptData =
         transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
       val events =
         listOf(
           transactionActivateEvent(),
           transactionAuthorizationRequestedEvent(),
           transactionAuthorizationCompletedEvent(),
           transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent)
+          notificationRequested)
           as List<TransactionEvent<Any>>
       val baseTransaction =
         reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
       val transactionId = TRANSACTION_ID
       val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
       Hooks.onOperatorDebug()
       given(checkpointer.success()).willReturn(Mono.empty())
       given(
@@ -350,7 +199,6 @@ class TransactionNotificationsRetryQueueConsumerTest {
       given(paymentGatewayClient.requestVPosRefund(any()))
         .willReturn(
           Mono.just(VposDeleteResponseDto().status(VposDeleteResponseDto.StatusEnum.CANCELLED)))
-
       given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
         .willReturnConsecutively(
           listOf(
@@ -359,25 +207,24 @@ class TransactionNotificationsRetryQueueConsumerTest {
             Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
             Mono.just(
               transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
+
       StepVerifier.create(
           transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(notificationErrorEvent).toBytes(), checkpointer))
+            BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
+            checkpointer))
         .expectNext()
         .verifyComplete()
       verify(checkpointer, times(1)).success()
       verify(transactionsEventStoreRepository, times(1))
         .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(3)).save(any())
       verify(transactionRefundRepository, times(2)).save(any())
       verify(paymentGatewayClient, times(1)).requestVPosRefund(any())
+      verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionsViewRepository, times(3)).save(any())
       verify(transactionUserReceiptRepository, times(1)).save(any())
+      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
       verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
       verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      val savedEvent = transactionUserReceiptCaptor.value
-      assertEquals(TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT, savedEvent.eventCode)
-      assertEquals(transactionUserReceiptData, savedEvent.data)
       val expectedStatuses =
         listOf(
           TransactionStatusDto.NOTIFIED_KO,
@@ -387,10 +234,256 @@ class TransactionNotificationsRetryQueueConsumerTest {
         listOf(
           TransactionEventCode.TRANSACTION_REFUND_REQUESTED_EVENT,
           TransactionEventCode.TRANSACTION_REFUNDED_EVENT)
+      assertEquals(
+        TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT,
+        TransactionEventCode.valueOf(transactionUserReceiptCaptor.value.eventCode))
+      assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.value.data)
       expectedEventCodes.forEachIndexed { index, eventCode ->
-        assertEquals(eventCode, transactionRefundEventStoreCaptor.allValues[index].eventCode)
         assertEquals(
-          TransactionStatusDto.NOTIFICATION_ERROR,
+          eventCode.toString(), transactionRefundEventStoreCaptor.allValues[index].eventCode)
+        assertEquals(
+          TransactionStatusDto.NOTIFICATION_REQUESTED,
+          transactionRefundEventStoreCaptor.allValues[index].data.statusBeforeRefunded)
+      }
+      expectedStatuses.forEachIndexed { index, transactionStatus ->
+        assertEquals(transactionStatus, transactionViewRepositoryCaptor.allValues[index].status)
+      }
+    }
+
+  @Test
+  fun `Should successfully send user email for send payment result outcome OK with legacy event`() =
+    runTest {
+      val transactionUserReceiptData =
+        transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      val document =
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
+        .willReturn(Flux.fromIterable(events))
+      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
+        .willReturn(NotificationEmailRequestDto())
+      given(notificationsServiceClient.sendNotificationEmail(any()))
+        .willReturn(Mono.just(NotificationEmailResponseDto().apply { outcome = "OK" }))
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturn(Mono.just(document))
+      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
+        Mono.just(it.arguments[0])
+      }
+      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+
+      StepVerifier.create(
+          transactionNotificationsRetryQueueConsumer.messageReceiver(
+            BinaryData.fromObject(notificationRequested).toBytes(), checkpointer))
+        .expectNext()
+        .verifyComplete()
+      verify(checkpointer, times(1)).success()
+      verify(transactionsEventStoreRepository, times(1))
+        .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
+      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
+      verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionsViewRepository, times(1)).save(any())
+      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionRefundRepository, times(0)).save(any())
+      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
+      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
+      assertEquals(TransactionStatusDto.NOTIFIED_OK, transactionViewRepositoryCaptor.value.status)
+      val savedEvent = transactionUserReceiptCaptor.value
+      assertEquals(
+        TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT,
+        TransactionEventCode.valueOf(savedEvent.eventCode))
+      assertEquals(transactionUserReceiptData, savedEvent.data)
+    }
+
+  @Test
+  fun `Should successfully send user email for send payment result outcome KO performing refund for transaction with legacy event`() =
+    runTest {
+      val transactionUserReceiptData =
+        transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      val document =
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
+        .willReturn(Flux.fromIterable(events))
+      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
+        .willReturn(NotificationEmailRequestDto())
+      given(notificationsServiceClient.sendNotificationEmail(any()))
+        .willReturn(Mono.just(NotificationEmailResponseDto().apply { outcome = "OK" }))
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturn(Mono.just(document))
+      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
+        Mono.just(it.arguments[0])
+      }
+      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+      given(transactionRefundRepository.save(capture(transactionRefundEventStoreCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+      given(paymentGatewayClient.requestVPosRefund(any()))
+        .willReturn(
+          Mono.just(VposDeleteResponseDto().status(VposDeleteResponseDto.StatusEnum.CANCELLED)))
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturnConsecutively(
+          listOf(
+            Mono.just(
+              transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())),
+            Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
+            Mono.just(
+              transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
+
+      StepVerifier.create(
+          transactionNotificationsRetryQueueConsumer.messageReceiver(
+            BinaryData.fromObject(notificationRequested).toBytes(), checkpointer))
+        .expectNext()
+        .verifyComplete()
+      verify(checkpointer, times(1)).success()
+      verify(transactionsEventStoreRepository, times(1))
+        .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
+      verify(transactionRefundRepository, times(2)).save(any())
+      verify(paymentGatewayClient, times(1)).requestVPosRefund(any())
+      verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionsViewRepository, times(3)).save(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
+      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
+      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
+      val expectedStatuses =
+        listOf(
+          TransactionStatusDto.NOTIFIED_KO,
+          TransactionStatusDto.REFUND_REQUESTED,
+          TransactionStatusDto.REFUNDED)
+      val expectedEventCodes =
+        listOf(
+          TransactionEventCode.TRANSACTION_REFUND_REQUESTED_EVENT,
+          TransactionEventCode.TRANSACTION_REFUNDED_EVENT)
+      assertEquals(
+        TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT,
+        TransactionEventCode.valueOf(transactionUserReceiptCaptor.value.eventCode))
+      assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.value.data)
+      expectedEventCodes.forEachIndexed { index, eventCode ->
+        assertEquals(
+          eventCode.toString(), transactionRefundEventStoreCaptor.allValues[index].eventCode)
+        assertEquals(
+          TransactionStatusDto.NOTIFICATION_REQUESTED,
+          transactionRefundEventStoreCaptor.allValues[index].data.statusBeforeRefunded)
+      }
+      expectedStatuses.forEachIndexed { index, transactionStatus ->
+        assertEquals(transactionStatus, transactionViewRepositoryCaptor.allValues[index].status)
+      }
+    }
+
+  @Test
+  fun `Should successfully send user email for send payment result outcome KO and enqueue refund retry event in case of error performing refund`() =
+    runTest {
+      val transactionUserReceiptData =
+        transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      val document =
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
+        .willReturn(Flux.fromIterable(events))
+      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
+        .willReturn(NotificationEmailRequestDto())
+      given(notificationsServiceClient.sendNotificationEmail(any()))
+        .willReturn(Mono.just(NotificationEmailResponseDto().apply { outcome = "OK" }))
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturn(Mono.just(document))
+      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
+        Mono.just(it.arguments[0])
+      }
+      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+      given(transactionRefundRepository.save(capture(transactionRefundEventStoreCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+      given(paymentGatewayClient.requestVPosRefund(any()))
+        .willReturn(Mono.error(RuntimeException("Error performing refunding")))
+      given(refundRetryService.enqueueRetryEvent(any(), any(), any())).willReturn(Mono.empty())
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturnConsecutively(
+          listOf(
+            Mono.just(
+              transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())),
+            Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
+            Mono.just(
+              transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
+
+      StepVerifier.create(
+          transactionNotificationsRetryQueueConsumer.messageReceiver(
+            BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
+            checkpointer))
+        .expectNext()
+        .verifyComplete()
+      verify(checkpointer, times(1)).success()
+      verify(transactionsEventStoreRepository, times(1))
+        .findByTransactionIdOrderByCreationDateAsc(TRANSACTION_ID)
+      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
+      verify(notificationRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionRefundRepository, times(2)).save(any())
+      verify(paymentGatewayClient, times(1)).requestVPosRefund(any())
+      verify(transactionsViewRepository, times(3)).save(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
+      verify(refundRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
+      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
+      val expectedStatuses =
+        listOf(
+          TransactionStatusDto.NOTIFIED_KO,
+          TransactionStatusDto.REFUND_REQUESTED,
+          TransactionStatusDto.REFUND_ERROR)
+      val expectedEventCodes =
+        listOf(
+          TransactionEventCode.TRANSACTION_REFUND_REQUESTED_EVENT,
+          TransactionEventCode.TRANSACTION_REFUND_ERROR_EVENT)
+      assertEquals(
+        TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT,
+        TransactionEventCode.valueOf(transactionUserReceiptCaptor.value.eventCode))
+      assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.value.data)
+      expectedEventCodes.forEachIndexed { index, eventCode ->
+        assertEquals(
+          eventCode.toString(), transactionRefundEventStoreCaptor.allValues[index].eventCode)
+        assertEquals(
+          TransactionStatusDto.NOTIFICATION_REQUESTED,
           transactionRefundEventStoreCaptor.allValues[index].data.statusBeforeRefunded)
       }
       expectedStatuses.forEachIndexed { index, transactionStatus ->
@@ -403,21 +496,20 @@ class TransactionNotificationsRetryQueueConsumerTest {
     runTest {
       val transactionUserReceiptData =
         transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
       val events =
         listOf(
           transactionActivateEvent(),
           transactionAuthorizationRequestedEvent(),
           transactionAuthorizationCompletedEvent(),
           transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent)
+          notificationRequested)
           as List<TransactionEvent<Any>>
       val baseTransaction =
         reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
       val transactionId = TRANSACTION_ID
       val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
       Hooks.onOperatorDebug()
       given(checkpointer.success()).willReturn(Mono.empty())
       given(
@@ -439,7 +531,7 @@ class TransactionNotificationsRetryQueueConsumerTest {
         .willReturn(Mono.empty())
       StepVerifier.create(
           transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(QueueEvent(notificationErrorEvent, MOCK_TRACING_INFO)).toBytes(),
+            BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
             checkpointer))
         .expectNext()
         .verifyComplete()
@@ -448,14 +540,20 @@ class TransactionNotificationsRetryQueueConsumerTest {
         .findByTransactionIdOrderByCreationDateAsc(transactionId)
       verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
       verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(0)).save(any())
+      verify(transactionsViewRepository, times(1)).save(any())
       verify(transactionRefundRepository, times(0)).save(any())
       verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
       verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
       verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
 
       assertEquals(0, retryCountCaptor.value)
+      assertEquals(
+        TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT,
+        TransactionEventCode.valueOf(transactionUserReceiptCaptor.value.eventCode))
+      assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.value.data)
+      assertEquals(
+        TransactionStatusDto.NOTIFICATION_ERROR, transactionViewRepositoryCaptor.value.status)
     }
 
   @Test
@@ -463,21 +561,20 @@ class TransactionNotificationsRetryQueueConsumerTest {
     runTest {
       val transactionUserReceiptData =
         transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
       val events =
         listOf(
           transactionActivateEvent(),
           transactionAuthorizationRequestedEvent(),
           transactionAuthorizationCompletedEvent(),
           transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent)
+          notificationRequested)
           as List<TransactionEvent<Any>>
       val baseTransaction =
         reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
       val transactionId = TRANSACTION_ID
       val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
       Hooks.onOperatorDebug()
       given(checkpointer.success()).willReturn(Mono.empty())
       given(
@@ -499,7 +596,7 @@ class TransactionNotificationsRetryQueueConsumerTest {
         .willReturn(Mono.empty())
       StepVerifier.create(
           transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(QueueEvent(notificationErrorEvent, MOCK_TRACING_INFO)).toBytes(),
+            BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
             checkpointer))
         .expectNext()
         .verifyComplete()
@@ -508,379 +605,21 @@ class TransactionNotificationsRetryQueueConsumerTest {
         .findByTransactionIdOrderByCreationDateAsc(transactionId)
       verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
       verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(0)).save(any())
+      verify(transactionsViewRepository, times(1)).save(any())
       verify(transactionRefundRepository, times(0)).save(any())
       verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
       verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
       verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
       assertEquals(0, retryCountCaptor.value)
-    }
-
-  @Test
-  fun `Should enqueue notification retry event for failure retry calling notification service send payment result outcome OK`() =
-    runTest {
-      val attempt = 1
-      val transactionUserReceiptData =
-        transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-      val notificationRetriedEvent = transactionUserReceiptAddRetriedEvent(attempt)
-      val events =
-        listOf(
-          transactionActivateEvent(),
-          transactionAuthorizationRequestedEvent(),
-          transactionAuthorizationCompletedEvent(),
-          transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent,
-          notificationRetriedEvent)
-          as List<TransactionEvent<Any>>
-      val baseTransaction =
-        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-      val transactionId = TRANSACTION_ID
-      val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-      Hooks.onOperatorDebug()
-      given(checkpointer.success()).willReturn(Mono.empty())
-      given(
-          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-            TRANSACTION_ID))
-        .willReturn(Flux.fromIterable(events))
-      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-        .willReturn(NotificationEmailRequestDto())
-      given(notificationsServiceClient.sendNotificationEmail(any()))
-        .willReturn(Mono.error(RuntimeException("Error calling notification service")))
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturn(Mono.just(document))
-      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-        Mono.just(it.arguments[0])
-      }
-      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), any()))
-        .willReturn(Mono.empty())
-      StepVerifier.create(
-          transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(QueueEvent(notificationRetriedEvent, MOCK_TRACING_INFO))
-              .toBytes(),
-            checkpointer))
-        .expectNext()
-        .verifyComplete()
-      verify(checkpointer, times(1)).success()
-      verify(transactionsEventStoreRepository, times(1))
-        .findByTransactionIdOrderByCreationDateAsc(transactionId)
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(0)).save(any())
-      verify(transactionRefundRepository, times(0)).save(any())
-      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
-      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      assertEquals(attempt, retryCountCaptor.value)
-    }
-
-  @Test
-  fun `Should enqueue notification retry event for failure retry calling notification service send payment result outcome OK with legacy event`() =
-    runTest {
-      val attempt = 1
-      val transactionUserReceiptData =
-        transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-      val notificationRetriedEvent = transactionUserReceiptAddRetriedEvent(attempt)
-      val events =
-        listOf(
-          transactionActivateEvent(),
-          transactionAuthorizationRequestedEvent(),
-          transactionAuthorizationCompletedEvent(),
-          transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent,
-          notificationRetriedEvent)
-          as List<TransactionEvent<Any>>
-      val baseTransaction =
-        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-      val transactionId = TRANSACTION_ID
-      val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-      Hooks.onOperatorDebug()
-      given(checkpointer.success()).willReturn(Mono.empty())
-      given(
-          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-            TRANSACTION_ID))
-        .willReturn(Flux.fromIterable(events))
-      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-        .willReturn(NotificationEmailRequestDto())
-      given(notificationsServiceClient.sendNotificationEmail(any()))
-        .willReturn(Mono.error(RuntimeException("Error calling notification service")))
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturn(Mono.just(document))
-      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-        Mono.just(it.arguments[0])
-      }
-      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), isNull()))
-        .willReturn(Mono.empty())
-      StepVerifier.create(
-          transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(notificationRetriedEvent).toBytes(), checkpointer))
-        .expectNext()
-        .verifyComplete()
-      verify(checkpointer, times(1)).success()
-      verify(transactionsEventStoreRepository, times(1))
-        .findByTransactionIdOrderByCreationDateAsc(transactionId)
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), isNull())
-      verify(transactionsViewRepository, times(0)).save(any())
-      verify(transactionRefundRepository, times(0)).save(any())
-      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
-      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), isNull())
-      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      assertEquals(attempt, retryCountCaptor.value)
-    }
-
-  @Test
-  fun `Should enqueue notification retry event for failure retry calling notification service send payment result outcome KO`() =
-    runTest {
-      val attempt = 1
-      val transactionUserReceiptData =
-        transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-      val notificationRetriedEvent = transactionUserReceiptAddRetriedEvent(attempt)
-      val events =
-        listOf(
-          transactionActivateEvent(),
-          transactionAuthorizationRequestedEvent(),
-          transactionAuthorizationCompletedEvent(),
-          transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent,
-          notificationRetriedEvent)
-          as List<TransactionEvent<Any>>
-      val baseTransaction =
-        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-      val transactionId = TRANSACTION_ID
-      val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-      Hooks.onOperatorDebug()
-      given(checkpointer.success()).willReturn(Mono.empty())
-      given(
-          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-            TRANSACTION_ID))
-        .willReturn(Flux.fromIterable(events))
-      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-        .willReturn(NotificationEmailRequestDto())
-      given(notificationsServiceClient.sendNotificationEmail(any()))
-        .willReturn(Mono.error(RuntimeException("Error calling notification service")))
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturn(Mono.just(document))
-      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-        Mono.just(it.arguments[0])
-      }
-      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), any()))
-        .willReturn(Mono.empty())
-
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturnConsecutively(
-          listOf(
-            Mono.just(
-              transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())),
-            Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
-            Mono.just(
-              transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
-      StepVerifier.create(
-          transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(QueueEvent(notificationRetriedEvent, MOCK_TRACING_INFO))
-              .toBytes(),
-            checkpointer))
-        .expectNext()
-        .verifyComplete()
-      verify(checkpointer, times(1)).success()
-      verify(transactionsEventStoreRepository, times(1))
-        .findByTransactionIdOrderByCreationDateAsc(transactionId)
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(0)).save(any())
-      verify(transactionRefundRepository, times(0)).save(any())
-      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
-      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      assertEquals(attempt, retryCountCaptor.value)
-    }
-
-  @Test
-  fun `Should not perform refund for no left attempts resending mail for send payment result OK`() =
-    runTest {
-      val attempt = 3
-      val transactionUserReceiptData =
-        transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-      val notificationRetriedEvent = transactionUserReceiptAddRetriedEvent(attempt)
-      val events =
-        listOf(
-          transactionActivateEvent(),
-          transactionAuthorizationRequestedEvent(),
-          transactionAuthorizationCompletedEvent(),
-          transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent,
-          notificationRetriedEvent)
-          as List<TransactionEvent<Any>>
-      val baseTransaction =
-        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-      val transactionId = TRANSACTION_ID
-      val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-      Hooks.onOperatorDebug()
-      given(checkpointer.success()).willReturn(Mono.empty())
-      given(
-          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-            TRANSACTION_ID))
-        .willReturn(Flux.fromIterable(events))
-      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-        .willReturn(NotificationEmailRequestDto())
-      given(notificationsServiceClient.sendNotificationEmail(any()))
-        .willReturn(Mono.error(RuntimeException("Error calling notification service")))
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturn(Mono.just(document))
-      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-        Mono.just(it.arguments[0])
-      }
-      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-
-      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), any()))
-        .willReturn(
-          Mono.error(
-            NoRetryAttemptsLeftException(
-              TransactionId(transactionId),
-              TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT)))
-
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturnConsecutively(
-          listOf(
-            Mono.just(
-              transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())),
-            Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
-            Mono.just(
-              transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
-
-      StepVerifier.create(
-          transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(QueueEvent(notificationRetriedEvent, MOCK_TRACING_INFO))
-              .toBytes(),
-            checkpointer))
-        .expectNext()
-        .verifyComplete()
-      verify(checkpointer, times(1)).success()
-      verify(transactionsEventStoreRepository, times(1))
-        .findByTransactionIdOrderByCreationDateAsc(transactionId)
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(0)).save(any())
-      verify(transactionRefundRepository, times(0)).save(any())
-      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
-      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      assertEquals(attempt, retryCountCaptor.value)
-    }
-
-  @Test
-  fun `Should perform refund for no left attempts resending mail for send payment result KO`() =
-    runTest {
-      val attempt = 3
-      val transactionUserReceiptData =
-        transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-      val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
-      val notificationRetriedEvent = transactionUserReceiptAddRetriedEvent(attempt)
-      val events =
-        listOf(
-          transactionActivateEvent(),
-          transactionAuthorizationRequestedEvent(),
-          transactionAuthorizationCompletedEvent(),
-          transactionClosedEvent(TransactionClosureData.Outcome.OK),
-          transactionUserReceiptRequestedEvent(transactionUserReceiptData),
-          notificationErrorEvent,
-          notificationRetriedEvent)
-          as List<TransactionEvent<Any>>
-      val baseTransaction =
-        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
-      val transactionId = TRANSACTION_ID
-      val document =
-        transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())
-      Hooks.onOperatorDebug()
-      given(checkpointer.success()).willReturn(Mono.empty())
-      given(
-          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-            TRANSACTION_ID))
-        .willReturn(Flux.fromIterable(events))
-      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
-        .willReturn(NotificationEmailRequestDto())
-      given(notificationsServiceClient.sendNotificationEmail(any()))
-        .willReturn(Mono.error(RuntimeException("Error calling notification service")))
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturn(Mono.just(document))
-      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
-        Mono.just(it.arguments[0])
-      }
-      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), any()))
-        .willReturn(
-          Mono.error(
-            NoRetryAttemptsLeftException(
-              TransactionId(transactionId),
-              TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT)))
-      given(transactionRefundRepository.save(capture(transactionRefundEventStoreCaptor)))
-        .willAnswer { Mono.just(it.arguments[0]) }
-      given(paymentGatewayClient.requestVPosRefund(any()))
-        .willReturn(
-          Mono.just(VposDeleteResponseDto().status(VposDeleteResponseDto.StatusEnum.CANCELLED)))
-
-      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
-        .willReturnConsecutively(
-          listOf(
-            Mono.just(
-              transactionDocument(TransactionStatusDto.NOTIFICATION_ERROR, ZonedDateTime.now())),
-            Mono.just(transactionDocument(TransactionStatusDto.EXPIRED, ZonedDateTime.now())),
-            Mono.just(
-              transactionDocument(TransactionStatusDto.REFUND_REQUESTED, ZonedDateTime.now()))))
-      StepVerifier.create(
-          transactionNotificationsRetryQueueConsumer.messageReceiver(
-            BinaryData.fromObject(QueueEvent(notificationRetriedEvent, MOCK_TRACING_INFO))
-              .toBytes(),
-            checkpointer))
-        .expectNext()
-        .verifyComplete()
-      verify(checkpointer, times(1)).success()
-      verify(transactionsEventStoreRepository, times(1))
-        .findByTransactionIdOrderByCreationDateAsc(transactionId)
-      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
-      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
-      verify(transactionsViewRepository, times(2)).save(any())
-      verify(transactionRefundRepository, times(2)).save(any())
-      verify(paymentGatewayClient, times(1)).requestVPosRefund(any())
-      verify(transactionUserReceiptRepository, times(0)).save(any())
-      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
-      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
-      assertEquals(attempt, retryCountCaptor.value)
       val expectedStatuses =
-        listOf(TransactionStatusDto.REFUND_REQUESTED, TransactionStatusDto.REFUNDED)
-      val expectedEventCodes =
         listOf(
-          TransactionEventCode.TRANSACTION_REFUND_REQUESTED_EVENT,
-          TransactionEventCode.TRANSACTION_REFUNDED_EVENT)
-      expectedEventCodes.forEachIndexed { index, eventCode ->
-        assertEquals(eventCode, transactionRefundEventStoreCaptor.allValues[index].eventCode)
-        assertEquals(
           TransactionStatusDto.NOTIFICATION_ERROR,
-          transactionRefundEventStoreCaptor.allValues[index].data.statusBeforeRefunded)
+        )
+      val expectedEventCodes = listOf(TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT)
+      expectedEventCodes.forEachIndexed { index, eventCode ->
+        assertEquals(eventCode.toString(), transactionUserReceiptCaptor.allValues[index].eventCode)
+        assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.allValues[index].data)
       }
       expectedStatuses.forEachIndexed { index, transactionStatus ->
         assertEquals(transactionStatus, transactionViewRepositoryCaptor.allValues[index].status)
@@ -888,10 +627,94 @@ class TransactionNotificationsRetryQueueConsumerTest {
     }
 
   @Test
+  fun `Should return mono error for failure enqueuing notification retry event send payment result OK`() =
+    runTest {
+      val transactionUserReceiptData =
+        transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      val document =
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
+            TRANSACTION_ID))
+        .willReturn(Flux.fromIterable(events))
+      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
+        .willReturn(NotificationEmailRequestDto())
+      given(notificationsServiceClient.sendNotificationEmail(any()))
+        .willReturn(Mono.error(RuntimeException("Error calling notification service")))
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturn(Mono.just(document))
+      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
+        Mono.just(it.arguments[0])
+      }
+      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), any()))
+        .willReturn(Mono.error(RuntimeException("Error enqueueing notification retry event")))
+      given(
+          deadLetterQueueAsyncClient.sendMessageWithResponse(any<BinaryData>(), any(), anyOrNull()))
+        .willReturn(queueSuccessfulResponse())
+
+      StepVerifier.create(
+          transactionNotificationsRetryQueueConsumer.messageReceiver(
+            BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
+            checkpointer))
+        .verifyComplete()
+      verify(checkpointer, times(1)).success()
+      verify(transactionsEventStoreRepository, times(1))
+        .findByTransactionIdOrderByCreationDateAsc(transactionId)
+      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
+      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionsViewRepository, times(1)).save(any())
+      verify(transactionRefundRepository, times(0)).save(any())
+      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
+      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
+      assertEquals(0, retryCountCaptor.value)
+      val expectedStatuses =
+        listOf(
+          TransactionStatusDto.NOTIFICATION_ERROR,
+        )
+      val expectedEventCodes = listOf(TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT)
+      expectedEventCodes.forEachIndexed { index, eventCode ->
+        assertEquals(eventCode.toString(), transactionUserReceiptCaptor.allValues[index].eventCode)
+        assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.allValues[index].data)
+      }
+      expectedStatuses.forEachIndexed { index, transactionStatus ->
+        assertEquals(transactionStatus, transactionViewRepositoryCaptor.allValues[index].status)
+      }
+      verify(deadLetterQueueAsyncClient, times(1))
+        .sendMessageWithResponse(
+          argThat<BinaryData> {
+            TransactionEventCode.valueOf(
+              this.toObject(
+                  object : TypeReference<QueueEvent<TransactionUserReceiptRequestedEvent>>() {})
+                .event
+                .eventCode) == TransactionEventCode.TRANSACTION_USER_RECEIPT_REQUESTED_EVENT
+          },
+          eq(Duration.ZERO),
+          eq(Duration.ofSeconds(DEAD_LETTER_QUEUE_TTL_SECONDS.toLong())))
+    }
+
+  @Test
   fun `Should not process event for wrong transaction status`() = runTest {
     val transactionUserReceiptData =
       transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-    val notificationErrorEvent = transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+    val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
     val events = listOf(transactionActivateEvent()) as List<TransactionEvent<Any>>
     val transactionId = TRANSACTION_ID
     Hooks.onOperatorDebug()
@@ -903,7 +726,7 @@ class TransactionNotificationsRetryQueueConsumerTest {
 
     StepVerifier.create(
         transactionNotificationsRetryQueueConsumer.messageReceiver(
-          BinaryData.fromObject(QueueEvent(notificationErrorEvent, MOCK_TRACING_INFO)).toBytes(),
+          BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
           checkpointer))
       .verifyComplete()
     verify(checkpointer, times(1)).success()
@@ -920,14 +743,162 @@ class TransactionNotificationsRetryQueueConsumerTest {
     verify(deadLetterQueueAsyncClient, times(1))
       .sendMessageWithResponse(
         argThat<BinaryData> {
-          this.toObject(
-              object : TypeReference<QueueEvent<TransactionUserReceiptAddErrorEvent>>() {})
-            .event
-            .eventCode == TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT
+          TransactionEventCode.valueOf(
+            this.toObject(
+                object : TypeReference<QueueEvent<TransactionUserReceiptRequestedEvent>>() {})
+              .event
+              .eventCode) == TransactionEventCode.TRANSACTION_USER_RECEIPT_REQUESTED_EVENT
         },
         eq(Duration.ZERO),
         eq(Duration.ofSeconds(DEAD_LETTER_QUEUE_TTL_SECONDS.toLong())))
   }
+
+  @Test
+  fun `Should set right value string to payee template name field when TransactionUserReceiptData receivingOfficeName is not null`() =
+    runTest {
+      val confidentialMailUtils: ConfidentialMailUtils = mock()
+      given(confidentialMailUtils.toEmail(any())).willReturn(Email("to@to.it"))
+      val userReceiptBuilder = UserReceiptMailBuilder(confidentialMailUtils)
+      val transactionUserReceiptData =
+        TransactionUserReceiptData(
+          TransactionUserReceiptData.Outcome.OK,
+          "it-IT",
+          PAYMENT_DATE,
+          "testValue",
+          "paymentDescription")
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
+        .willReturn(Flux.fromIterable(events))
+
+      val notificationEmailRequestDto =
+        userReceiptBuilder.buildNotificationEmailRequestDto(baseTransaction)
+      assertEquals(
+        "testValue",
+        (notificationEmailRequestDto.parameters as SuccessTemplate)
+          .cart
+          .items
+          .filter { i -> i.payee != null }[0]
+          .payee
+          .name)
+    }
+
+  @Test
+  fun `Should set empty string to payee template name field when TransactionUserReceiptData receivingOfficeName is null`() =
+    runTest {
+      val confidentialMailUtils: ConfidentialMailUtils = mock()
+      given(confidentialMailUtils.toEmail(any())).willReturn(Email("to@to.it"))
+      val userReceiptBuilder = UserReceiptMailBuilder(confidentialMailUtils)
+      val transactionUserReceiptData =
+        TransactionUserReceiptData(
+          TransactionUserReceiptData.Outcome.OK, "it-IT", PAYMENT_DATE, null, "paymentDescription")
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(transactionId))
+        .willReturn(Flux.fromIterable(events))
+
+      val notificationEmailRequestDto =
+        userReceiptBuilder.buildNotificationEmailRequestDto(baseTransaction)
+      assertEquals(
+        "",
+        (notificationEmailRequestDto.parameters as SuccessTemplate)
+          .cart
+          .items
+          .filter { i -> i.payee != null }[0]
+          .payee
+          .name)
+    }
+
+  @Test
+  fun `Should not process event for transaction with invalid send payment result outcome`() =
+    runTest {
+      val transactionUserReceiptData =
+        transactionUserReceiptData(TransactionUserReceiptData.Outcome.NOT_RECEIVED)
+      val notificationRequested = transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+      val events =
+        listOf(
+          transactionActivateEvent(),
+          transactionAuthorizationRequestedEvent(),
+          transactionAuthorizationCompletedEvent(),
+          transactionClosedEvent(TransactionClosureData.Outcome.OK),
+          notificationRequested)
+          as List<TransactionEvent<Any>>
+      val baseTransaction =
+        reduceEvents(*events.toTypedArray()) as BaseTransactionWithRequestedUserReceipt
+      val transactionId = TRANSACTION_ID
+      val document =
+        transactionDocument(TransactionStatusDto.NOTIFICATION_REQUESTED, ZonedDateTime.now())
+      Hooks.onOperatorDebug()
+      given(checkpointer.success()).willReturn(Mono.empty())
+      given(
+          transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
+            TRANSACTION_ID))
+        .willReturn(Flux.fromIterable(events))
+      given(userReceiptMailBuilder.buildNotificationEmailRequestDto(baseTransaction))
+        .willReturn(NotificationEmailRequestDto())
+      given(notificationsServiceClient.sendNotificationEmail(any()))
+        .willReturn(Mono.just(NotificationEmailResponseDto().apply { outcome = "OK" }))
+      given(transactionsViewRepository.findByTransactionId(TRANSACTION_ID))
+        .willReturn(Mono.just(document))
+      given(transactionsViewRepository.save(capture(transactionViewRepositoryCaptor))).willAnswer {
+        Mono.just(it.arguments[0])
+      }
+      given(transactionUserReceiptRepository.save(capture(transactionUserReceiptCaptor)))
+        .willAnswer { Mono.just(it.arguments[0]) }
+      given(notificationRetryService.enqueueRetryEvent(any(), capture(retryCountCaptor), any()))
+        .willReturn(Mono.empty())
+      StepVerifier.create(
+          transactionNotificationsRetryQueueConsumer.messageReceiver(
+            BinaryData.fromObject(QueueEvent(notificationRequested, MOCK_TRACING_INFO)).toBytes(),
+            checkpointer))
+        .expectNext()
+        .verifyComplete()
+      verify(checkpointer, times(1)).success()
+      verify(transactionsEventStoreRepository, times(1))
+        .findByTransactionIdOrderByCreationDateAsc(transactionId)
+      verify(notificationsServiceClient, times(1)).sendNotificationEmail(any())
+      verify(notificationRetryService, times(1)).enqueueRetryEvent(any(), any(), any())
+      verify(transactionsViewRepository, times(1)).save(any())
+      verify(transactionRefundRepository, times(0)).save(any())
+      verify(paymentGatewayClient, times(0)).requestVPosRefund(any())
+      verify(transactionUserReceiptRepository, times(1)).save(any())
+      verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
+      verify(userReceiptMailBuilder, times(1)).buildNotificationEmailRequestDto(baseTransaction)
+
+      assertEquals(0, retryCountCaptor.value)
+      assertEquals(
+        TransactionEventCode.TRANSACTION_ADD_USER_RECEIPT_ERROR_EVENT,
+        TransactionEventCode.valueOf(transactionUserReceiptCaptor.value.eventCode))
+      assertEquals(transactionUserReceiptData, transactionUserReceiptCaptor.value.data)
+      assertEquals(
+        TransactionStatusDto.NOTIFICATION_ERROR, transactionViewRepositoryCaptor.value.status)
+    }
 
   @Test
   fun `consumer write event to dead letter queue for un-parsable event`() = runTest {
