+++ src/test/kotlin/it/pagopa/ecommerce/eventdispatcher/queues/v1/TransactionRefundRetryQueueConsumerTest.kt
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.eventdispatcher.queues
+package it.pagopa.ecommerce.eventdispatcher.queues.v1
 
 import com.azure.core.util.BinaryData
 import com.azure.core.util.serializer.TypeReference
@@ -14,7 +14,7 @@ import it.pagopa.ecommerce.commons.v1.TransactionTestUtils
 import it.pagopa.ecommerce.eventdispatcher.client.PaymentGatewayClient
 import it.pagopa.ecommerce.eventdispatcher.repositories.TransactionsEventStoreRepository
 import it.pagopa.ecommerce.eventdispatcher.repositories.TransactionsViewRepository
-import it.pagopa.ecommerce.eventdispatcher.services.eventretry.RefundRetryService
+import it.pagopa.ecommerce.eventdispatcher.services.eventretry.v1.RefundRetryService
 import it.pagopa.ecommerce.eventdispatcher.utils.DEAD_LETTER_QUEUE_TTL_SECONDS
 import it.pagopa.ecommerce.eventdispatcher.utils.queueSuccessfulResponse
 import it.pagopa.generated.ecommerce.gateway.v1.dto.VposDeleteResponseDto
@@ -142,7 +142,7 @@ class TransactionRefundRetryQueueConsumerTest {
       "Unexpected view status")
     assertEquals(
       TransactionEventCode.TRANSACTION_REFUNDED_EVENT,
-      transactionRefundEventStoreCaptor.value.eventCode,
+      TransactionEventCode.valueOf(transactionRefundEventStoreCaptor.value.eventCode),
       "Unexpected event code")
     verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
   }
@@ -220,7 +220,7 @@ class TransactionRefundRetryQueueConsumerTest {
       assertEquals(TransactionStatusDto.REFUND_ERROR, transactionViewRepositoryCaptor.value.status)
       assertEquals(
         TransactionEventCode.TRANSACTION_REFUND_ERROR_EVENT,
-        transactionRefundEventStoreCaptor.value.eventCode)
+        TransactionEventCode.valueOf(transactionRefundEventStoreCaptor.value.eventCode))
     }
 
   @Test
@@ -290,7 +290,7 @@ class TransactionRefundRetryQueueConsumerTest {
       "Unexpected view status")
     assertEquals(
       TransactionEventCode.TRANSACTION_REFUNDED_EVENT,
-      transactionRefundEventStoreCaptor.value.eventCode,
+      TransactionEventCode.valueOf(transactionRefundEventStoreCaptor.value.eventCode),
       "Unexpected event code")
     verify(refundRetryService, times(0)).enqueueRetryEvent(any(), any(), any())
   }
@@ -367,7 +367,7 @@ class TransactionRefundRetryQueueConsumerTest {
       assertEquals(TransactionStatusDto.REFUND_ERROR, transactionViewRepositoryCaptor.value.status)
       assertEquals(
         TransactionEventCode.TRANSACTION_REFUND_ERROR_EVENT,
-        transactionRefundEventStoreCaptor.value.eventCode)
+        TransactionEventCode.valueOf(transactionRefundEventStoreCaptor.value.eventCode))
     }
 
   @Test
@@ -490,9 +490,10 @@ class TransactionRefundRetryQueueConsumerTest {
       verify(deadLetterQueueAsyncClient, times(1))
         .sendMessageWithResponse(
           argThat<BinaryData> {
-            this.toObject(object : TypeReference<QueueEvent<TransactionRefundRetriedEvent>>() {})
-              .event
-              .eventCode == TransactionEventCode.TRANSACTION_REFUND_RETRIED_EVENT
+            TransactionEventCode.valueOf(
+              this.toObject(object : TypeReference<QueueEvent<TransactionRefundRetriedEvent>>() {})
+                .event
+                .eventCode) == TransactionEventCode.TRANSACTION_REFUND_RETRIED_EVENT
           },
           eq(Duration.ZERO),
           eq(Duration.ofSeconds(DEAD_LETTER_QUEUE_TTL_SECONDS.toLong())))
