+++ src/test/kotlin/it/pagopa/ecommerce/eventdispatcher/services/NodeServiceTests.kt
@@ -2,7 +2,7 @@ package it.pagopa.ecommerce.eventdispatcher.services
 
 import it.pagopa.ecommerce.commons.documents.v1.Transaction
 import it.pagopa.ecommerce.commons.documents.v1.TransactionEvent
-import it.pagopa.ecommerce.commons.domain.v1.TransactionId
+import it.pagopa.ecommerce.commons.domain.TransactionId
 import it.pagopa.ecommerce.commons.generated.server.model.AuthorizationResultDto
 import it.pagopa.ecommerce.commons.utils.EuroUtils
 import it.pagopa.ecommerce.commons.v1.TransactionTestUtils.*
